Move declaration of xsd:duration to owlready2 submodule
