+++ owlapy/owlready2/_base.py
@@ -4,6 +4,8 @@ from logging import warning
 from typing import Iterable, Set, Final, cast
 
 import owlready2
+from pandas import Timedelta
+from owlready2 import declare_datatype
 
 from owlapy import namespaces
 from owlapy.model import OWLOntologyManager, OWLOntology, OWLClass, OWLDataProperty, OWLObjectProperty, \
@@ -12,6 +14,18 @@ from owlapy.model import OWLOntologyManager, OWLOntology, OWLClass, OWLDataPrope
     IRI, OWLObjectInverseOf \
     # OWLObjectSomeValuesFrom, OWLProperty, \
 from owlapy.owlready2.utils import ToOwlready2
+
+
+def _parse_duration_datatype(literal: str):
+    return Timedelta(literal)
+
+
+def _unparse_duration_datatype(literal: Timedelta):
+    return literal.isoformat()
+
+
+declare_datatype(Timedelta, "http://www.w3.org/2001/XMLSchema#duration",
+                 _parse_duration_datatype, _unparse_duration_datatype)
 
 
 class BaseReasoner_Owlready2(Enum):
